// src/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  agent: typeof agent;
  auth: typeof auth;
  chat: typeof chat;
  company: typeof company;
  document: typeof document;
  http: typeof http;
  parser: typeof parser;
}>;
// ...
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;